requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.5
django-environ==0.4.5
django-heroku==0.3.1
gunicorn==20.1.0
ibm-cloud-sdk-core==3.10.1
ibm-watson==5.2.2
idna==3.2
psycopg2==2.7.7
PyAudio==0.2.11
PyJWT==2.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.1.0
whitenoise==5.3.0
